Fixed Something, but having many issues
I started with a repository that I made incorrectly somehow, so I made a
new repository the correct way. I tried to use this one but I couldn't
get it to work. I don't know what repository I'm using or committing to
anymore, so I have no way to know if anything is right or not and I
didn't have time to visit the resource center before and the TA left and
I don't know what to do.