I don't even know at this point
This whole this is just stupid and I have no way to test any of it
because the files won't save or open or run in Java so I changed the log
from "n" to "square" in the hopes that the mod operator wasn't checking
enough values. I have no way to check if this works or not, this is the
last thing I can possibly think of with no way to actually test the
code. I give up on this assignment for I have no other options at this
time. This course teaches us the languages as if we have no prior
knowledge, but it does not act as though we have no knowledge of
computers even though it should.